chore: formatted files

+++ src/Multi-Sig-Wallet/interfaces/IMultiSigWallet.sol
@@ -3,11 +3,11 @@ pragma solidity 0.8.20;
 
 interface IMultiSigWallet {
     // events
-    event Deposit(address indexed sender, uint amount);
-    event Submit(uint indexed txId);
-    event Approve(address indexed owner, uint indexed txId);
-    event Revoke(address indexed owner, uint indexed txId);
-    event Execute(uint indexed txId);
+    event Deposit(address indexed sender, uint256 amount);
+    event Submit(uint256 indexed txId);
+    event Approve(address indexed owner, uint256 indexed txId);
+    event Revoke(address indexed owner, uint256 indexed txId);
+    event Execute(uint256 indexed txId);
 
     // errors
     error OwnerIsRequired();
